modules/person_api/requirements.txt
aniso8601==7.0.0
attrs==19.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cirq==0.11.1
cirq-core==0.11.1
cirq-google==0.11.1
Click==7.0
cycler==0.10.0
Flask==1.1.1
flask-accepts==0.10.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
flask-restplus==0.12.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
GeoAlchemy2==0.8.4
google-api-core==1.31.1
google-auth==1.34.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-tools==1.42.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.0.2
kafka==1.3.5
kafka-python==2.0.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.2
mpmath==1.2.1
networkx==2.6.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
Pillow==8.3.1
protobuf==3.17.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
Shapely==1.8.0
six==1.15.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.19
sympy==1.8
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==0.16.1
